🔨🤖 move checkCanTagGdocType to its own module
Fixes the react(only-export-components) oxlint error in GdocsTags.tsx.
